docs: l10n and development workflow
The localization lifecycle is different from the development lifecycle
and uses different tools. The proposed workflow allows developers and
translators to work in parallel without blocking each other provided
the developers do not modified the localization files by other means.

Weblate needs to be able to push commits somewhere. Granting write
permissions to the official SecureDrop repository does not sound like
a sane thing to do. Instead, a dedicated bot user was created and owns
a fork of the SecureDrop repository, at http://lab.securedrop.club.

Weblate pushes commits with new translations to the i18n branch of
this repository. These commits are collected from time to time and
proposed as a pull request for merge into the develop branch.

In parallel developers modify the develop branch of the official
repository and may modify, remove or insert new strings to be
translated. From time to time a developer needs to:

* merge the develop branch into the i18n branch
* run ./manage.py translate --extract-update

Merging develop into i18n will never generate conflicts because
developers never modify the files same files as weblate.

To keep the develop branch as clean as possible, the commits created
by weblate are cherry-picked. Although it would be possible to merge
the i18n branch back into develop, the merge commits unique to the
i18n branch would also be added to the develop branch and make it more
difficult to read.